+++ ReleaseNotes.md
@@ -1,6 +1,18 @@
 Release notes for Bodoconsult.App
 ==========================
 
+# Bodoconsult.App.Abstractions
+
+## 1.0.5 
+
+Extracted abstractions to Bodoconsult.App.Abstractions
+
+## 1.0.6
+
+Added some properties for I18N to IAppGlobals
+
+
+
 # Bodoconsult.App
 
 ## 1.0.0
@@ -21,6 +33,16 @@ Bugfixes
 
 Migrated CredentialManager implementation to Bodoconsult.App.Windows package
 
+## 1.0.5 
+
+Extracted abstractions to Bodoconsult.App.Abstractions
+
+## 1.0.6
+
+Using Bodoconsult.App.Abstractions 1.0.6
+
+
+
 # Bodoconsult.App.Windows
 
 ## 1.0.0
@@ -30,3 +52,33 @@ New package
 ## 1.0.4
 
 Moved AD, DNS and DHCP related classes to Bodoconsult.Network repo and removed dependencies Microsoft.Management.Infrastructure and Microsoft.Windows.Compatibility to get a smaller footprint for package
+
+## 1.0.5
+
+Migrated interfaces to Bodoconsult.App.Abstractions
+
+## 1.0.6
+
+Using Bodoconsult.App.Abstractions 1.0.6
+
+
+# Bodoconsult.App.Wpf
+
+## 1.0.5
+
+First draft
+
+## 1.0.6
+
+Bugfixing and improvements to central classes
+
+
+# Bodoconsult.App.WinForms
+
+## 1.0.5 
+
+Extracted abstractions to Bodoconsult.App.Abstractions
+
+## 1.0.6
+
+Using Bodoconsult.App.Abstractions 1.0.6